requirements.txt
Flask==2.3.3
torch==2.0.1
torchvision==0.15.2
opencv-python-headless==4.8.0.76
Pillow==10.0.0
requests==2.31.0
numpy==1.24.3
pandas==2.0.3
ultralytics==8.0.20
gunicorn==21.2.0